Some corrections and code refactoring
Separate file for "enums"
